requirements.txt
annotated-types
anyio
certifi
click
colorama
fastapi
greenlet
gunicorn
h11
httpcore
httpx
idna
iniconfig
packaging
passlib[bcrypt]
pluggy
pydantic
pydantic_core
pyjwt
pytest
python-dotenv
python-multipart
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn